Action Chunk Prediction/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
einops>=0.7.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
opencv-python>=4.9.0
